Add test and move test_slice_reader to the appropriate module.
