+++ src/libstd/io/cursor.rs
@@ -259,21 +259,23 @@ mod tests {
     }
 
     #[test]
-    fn test_slice_reader() {
-        let in_buf = vec![0, 1, 2, 3, 4, 5, 6, 7];
-        let mut reader = &mut in_buf.as_slice();
+    fn test_boxed_reader() {
+        let mut reader = Cursor::new(
+            vec![0, 1, 2, 3, 4, 5, 6, 7].into_boxed_slice(),
+        );
         let mut buf = [];
         assert_eq!(reader.read(&mut buf), Ok(0));
+        assert_eq!(reader.position(), 0);
         let mut buf = [0];
         assert_eq!(reader.read(&mut buf), Ok(1));
-        assert_eq!(reader.len(), 7);
+        assert_eq!(reader.position(), 1);
         let b: &[_] = &[0];
-        assert_eq!(buf.as_slice(), b);
+        assert_eq!(buf, b);
         let mut buf = [0; 4];
         assert_eq!(reader.read(&mut buf), Ok(4));
-        assert_eq!(reader.len(), 3);
+        assert_eq!(reader.position(), 5);
         let b: &[_] = &[1, 2, 3, 4];
-        assert_eq!(buf.as_slice(), b);
+        assert_eq!(buf, b);
         assert_eq!(reader.read(&mut buf), Ok(3));
         let b: &[_] = &[5, 6, 7];
         assert_eq!(&buf[..3], b);
